+++ app/src/main/kotlin/io/github/maxlyth/hapaneld/PaneldService.kt
@@ -206,6 +206,7 @@ class PaneldService : Service() {
         scope.launch {
             io.github.maxlyth.hapaneld.http.PerfReader.dashboardPkg = dashboardTarget()
             io.github.maxlyth.hapaneld.http.PerfReader.enabled = config.instrumentationEnabled
+            io.github.maxlyth.hapaneld.sensors.SensorTrace.enabled = config.instrumentationEnabled
             io.github.maxlyth.hapaneld.http.PerfReader.start(scope)
             server.start()
             mdns.start()